Added new hsobject and physicsobject hold attributes
Added velocity override and repositioning to HSObjectHold. Added gravity
ignorance to PhysicsObjectHold

+++ hsobject.h
@@ -187,7 +187,7 @@ public:
 	HSVect2D prevPos; //the object's previous position
 	HSVect2D vel; //the object's velocity
 	HSVect2D prevVel; //the object's previous velocity
-	int depth; //how far into the background (positive) or forground (negative) the object should be considered. effects zoom and parallax
+	float depth; //how far into the background (positive) or forground (negative) the object should be considered. effects zoom and parallax
 	HSVect2D reposition; //a shifting of position to be applied during the update phase
 	bool overwriteVelocity; //whether or not to overwrite the current velocity with the value in the below variable
 	HSVect2D holdVelocity; //velocity to apply in the update phase, if the above boolean is true
